Remove deprecated Near Cache modes
